# docs/swagger.yaml
basePath: /
definitions:
  AuthLoginRequest:
    properties:
      email:
        example: admin@example.com
        maxLength: 255
        type: string
      password:
        example: password123
        maxLength: 255
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  AuthRefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  AuthTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      access_token_expired_at:
        example: "2006-01-02T03:04:05+07:00"
        format: YYYY-MM-DDThh:mm:ssZ
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      refresh_token_expired_at:
        example: "2006-01-02T03:04:05+07:00"
        format: YYYY-MM-DDThh:mm:ssZ
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  Response:
    properties:
      data: {}
    type: object
  SuccessResponse:
    properties:
      message:
        example: OK
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Auction Service REST API
  title: Auction Service API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body Request
        in: body
        name: AuthLoginRequest
        required: true
        schema:
          $ref: '#/definitions/AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/AuthTokenResponse'
              type: object
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body Request
        in: body
        name: AuthRefreshTokenRequest
        required: true
        schema:
          $ref: '#/definitions/AuthRefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/AuthTokenResponse'
              type: object
      summary: Refresh Token
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"